Remove Redundant code
Removed reused code in plotter.py and time_plotter.py, updated readme.
